# opt: add support for reverse scans