--- a/pkg/sql/opt/memo/private_storage.go
+++ b/pkg/sql/opt/memo/private_storage.go
@@ -207,6 +207,11 @@ func (ps *privateStorage) internScanOpDef(def *ScanOpDef) PrivateID {
 	// The below code is carefully constructed to not allocate in the case where
 	// the value is already in the map. Be careful when modifying.
 	ps.keyBuf.Reset()
+	if def.Reverse {
+		ps.keyBuf.WriteByte(1)
+	} else {
+		ps.keyBuf.WriteByte(0)
+	}
 	ps.keyBuf.writeUvarint(uint64(def.Table))
 	ps.keyBuf.writeUvarint(uint64(def.Index))