--- a/pkg/sql/opt/norm/testdata/rules/combo
+++ b/pkg/sql/opt/norm/testdata/rules/combo
@@ -293,15 +293,10 @@ GenerateIndexScans (higher cost)
          │    ├── key: (1)
          │    ├── fd: (1)-->(2,4)
   -      │    ├── scan a
-  +      │    ├── index-join a
+  +      │    ├── scan a,rev
          │    │    ├── columns: k:1(int!null) i:2(int) s:4(string)
          │    │    ├── key: (1)
-  -      │    │    └── fd: (1)-->(2,4)
-  +      │    │    ├── fd: (1)-->(2,4)
-  +      │    │    └── scan a@secondary
-  +      │    │         ├── columns: k:1(int!null) s:4(string)
-  +      │    │         ├── key: (1)
-  +      │    │         └── fd: (1)-->(4)
+         │    │    └── fd: (1)-->(2,4)
          │    └── filters [type=bool, outer=(2), constraints=(/2: (/NULL - ])]
          │         └── a.i = (10 - 1) [type=bool, outer=(2), constraints=(/2: (/NULL - ])]
          ├── scan xy
@@ -310,11 +305,45 @@ GenerateIndexScans (higher cost)
          └── filters [type=bool, outer=(1,6), constraints=(/1: (/NULL - ]; /6: (/NULL - ]), fd=(1)==(6), (6)==(1)]
               └── a.k = xy.x [type=bool, outer=(1,6), constraints=(/1: (/NULL - ]; /6: (/NULL - ])]
 --------------------------------------------------------------------------------
+GenerateIndexScans (no changes)
+--------------------------------------------------------------------------------
+--------------------------------------------------------------------------------
+ConstrainScan (no changes)
+--------------------------------------------------------------------------------
+--------------------------------------------------------------------------------
 ConstrainScan (no changes)
 --------------------------------------------------------------------------------
 --------------------------------------------------------------------------------
 ConstrainIndexJoinScan (no changes)
 --------------------------------------------------------------------------------
+--------------------------------------------------------------------------------
+ConstrainIndexJoinScan (no changes)
+--------------------------------------------------------------------------------
+--------------------------------------------------------------------------------
+GenerateIndexScans (higher cost)
+--------------------------------------------------------------------------------
+   project
+    ├── columns: s:4(string)
+    └── inner-join
+         ├── columns: k:1(int!null) i:2(int!null) s:4(string) x:6(int!null)
+         ├── key: (6)
+         ├── fd: (1)-->(2,4), (1)==(6), (6)==(1)
+         ├── select
+         │    ├── columns: k:1(int!null) i:2(int!null) s:4(string)
+         │    ├── key: (1)
+         │    ├── fd: (1)-->(2,4)
+         │    ├── scan a
+         │    │    ├── columns: k:1(int!null) i:2(int) s:4(string)
+         │    │    ├── key: (1)
+         │    │    └── fd: (1)-->(2,4)
+         │    └── filters [type=bool, outer=(2), constraints=(/2: (/NULL - ])]
+         │         └── a.i = (10 - 1) [type=bool, outer=(2), constraints=(/2: (/NULL - ])]
+  -      ├── scan xy
+  +      ├── scan xy,rev
+         │    ├── columns: x:6(int!null)
+         │    └── key: (6)
+         └── filters [type=bool, outer=(1,6), constraints=(/1: (/NULL - ]; /6: (/NULL - ]), fd=(1)==(6), (6)==(1)]
+              └── a.k = xy.x [type=bool, outer=(1,6), constraints=(/1: (/NULL - ]; /6: (/NULL - ])]
 --------------------------------------------------------------------------------
 GenerateIndexScans (no changes)
 --------------------------------------------------------------------------------
@@ -508,6 +537,12 @@ GenerateIndexScans
               ├── a.s = 'foo' [type=bool, outer=(4), constraints=(/4: [/'foo' - /'foo']; tight)]
               └── a.f > 100.0 [type=bool, outer=(3), constraints=(/3: [/100.00000000000001 - ]; tight)]
 --------------------------------------------------------------------------------
+GenerateIndexScans (no changes)
+--------------------------------------------------------------------------------
+--------------------------------------------------------------------------------
+ConstrainScan (no changes)
+--------------------------------------------------------------------------------
+--------------------------------------------------------------------------------
 ConstrainScan (no changes)
 --------------------------------------------------------------------------------
 ================================================================================
@@ -558,6 +593,19 @@ SimplifyFilters
   -           ├── true [type=bool]
   -           └── true [type=bool]
   +      └── fd: ()-->(4), (1)-->(3), (3)-->(1)
+--------------------------------------------------------------------------------
+ConstrainScan (higher cost)
+--------------------------------------------------------------------------------
+   project
+    ├── columns: s:4(string!null) k:1(int!null)
+    ├── key: (1)
+    ├── fd: ()-->(4)
+  - └── scan a@secondary
+  + └── scan a@secondary,rev
+         ├── columns: k:1(int!null) f:3(float!null) s:4(string!null)
+         ├── constraint: /-4/3: [/'foo'/100.00000000000001 - /'foo']
+         ├── key: (1)
+         └── fd: ()-->(4), (1)-->(3), (3)-->(1)
 ================================================================================
 Final best expression
   Cost: 0.51
@@ -771,16 +819,32 @@ GenerateIndexScans (higher cost)
     ├── key: (1)
     ├── fd: (1)-->(2-5), (3,4)~~>(1,2,5)
   - ├── scan a
-  + ├── index-join a
+  + ├── scan a,rev
     │    ├── columns: k:1(int!null) i:2(int) f:3(float) s:4(string) j:5(jsonb)
     │    ├── key: (1)
-  - │    └── fd: (1)-->(2-5), (3,4)~~>(1,2,5)
-  + │    ├── fd: (1)-->(2-5), (3,4)~~>(1,2,5)
-  + │    └── scan a@secondary
-  + │         ├── columns: k:1(int!null) f:3(float) s:4(string) j:5(jsonb)
-  + │         ├── key: (1)
-  + │         └── fd: (1)-->(3-5), (3,4)~~>(1,5)
+    │    └── fd: (1)-->(2-5), (3,4)~~>(1,2,5)
     ├── scan xy
+    │    ├── columns: x:6(int!null) y:7(int)
+    │    ├── key: (6)
+    │    └── fd: (6)-->(7)
+    └── filters [type=bool, outer=(2,7), constraints=(/2: (/NULL - ]; /7: (/NULL - ]), fd=(2)==(7), (7)==(2)]
+         └── xy.y = a.i [type=bool, outer=(2,7), constraints=(/2: (/NULL - ]; /7: (/NULL - ])]
+--------------------------------------------------------------------------------
+GenerateIndexScans (no changes)
+--------------------------------------------------------------------------------
+--------------------------------------------------------------------------------
+GenerateIndexScans (higher cost)
+--------------------------------------------------------------------------------
+   semi-join
+    ├── columns: k:1(int!null) i:2(int) f:3(float) s:4(string) j:5(jsonb)
+    ├── key: (1)
+    ├── fd: (1)-->(2-5), (3,4)~~>(1,2,5)
+    ├── scan a
+    │    ├── columns: k:1(int!null) i:2(int) f:3(float) s:4(string) j:5(jsonb)
+    │    ├── key: (1)
+    │    └── fd: (1)-->(2-5), (3,4)~~>(1,2,5)
+  - ├── scan xy
+  + ├── scan xy,rev
     │    ├── columns: x:6(int!null) y:7(int)
     │    ├── key: (6)
     │    └── fd: (6)-->(7)
@@ -1597,6 +1661,40 @@ InlineProjectInProject
   + └── projections [outer=(1,10)]
   +      └── CASE WHEN bool_or AND (xy.x IS NOT NULL) THEN true WHEN bool_or IS NULL THEN false END [type=bool, outer=(1,10)]
 --------------------------------------------------------------------------------
+GenerateIndexScans (higher cost)
+--------------------------------------------------------------------------------
+   project
+    ├── columns: r:8(bool)
+    ├── group-by
+    │    ├── columns: x:1(int!null) bool_or:10(bool)
+    │    ├── grouping columns: x:1(int!null)
+    │    ├── key: (1)
+    │    ├── fd: (1)-->(10)
+    │    ├── left-join
+    │    │    ├── columns: x:1(int!null) k:3(int) notnull:9(bool)
+    │    │    ├── key: (1,3)
+    │    │    ├── fd: (3)-->(9)
+  - │    │    ├── scan xy
+  + │    │    ├── scan xy,rev
+    │    │    │    ├── columns: x:1(int!null)
+    │    │    │    └── key: (1)
+    │    │    ├── project
+    │    │    │    ├── columns: notnull:9(bool) k:3(int!null)
+    │    │    │    ├── key: (3)
+    │    │    │    ├── fd: (3)-->(9)
+    │    │    │    ├── scan a
+    │    │    │    │    ├── columns: k:3(int!null)
+    │    │    │    │    └── key: (3)
+    │    │    │    └── projections [outer=(3)]
+    │    │    │         └── a.k IS NOT NULL [type=bool, outer=(3)]
+    │    │    └── filters [type=bool, outer=(1,3)]
+    │    │         └── (xy.x = a.k) IS NOT false [type=bool, outer=(1,3)]
+    │    └── aggregations [outer=(9)]
+    │         └── bool-or [type=bool, outer=(9)]
+    │              └── variable: notnull [type=bool, outer=(9)]
+    └── projections [outer=(1,10)]
+         └── CASE WHEN bool_or AND (xy.x IS NOT NULL) THEN true WHEN bool_or IS NULL THEN false END [type=bool, outer=(1,10)]
+--------------------------------------------------------------------------------
 GenerateIndexScans (no changes)
 --------------------------------------------------------------------------------
 ================================================================================
@@ -1634,6 +1732,9 @@ GenerateIndexScans
     │              └── variable: notnull [type=bool, outer=(9)]
     └── projections [outer=(1,10)]
          └── CASE WHEN bool_or AND (xy.x IS NOT NULL) THEN true WHEN bool_or IS NULL THEN false END [type=bool, outer=(1,10)]
+--------------------------------------------------------------------------------
+GenerateIndexScans (no changes)
+--------------------------------------------------------------------------------
 --------------------------------------------------------------------------------
 CommuteLeftJoin (higher cost)
 --------------------------------------------------------------------------------