--- a/pkg/sql/opt/xform/testdata/physprops/ordering
+++ b/pkg/sql/opt/xform/testdata/physprops/ordering
@@ -182,7 +182,7 @@ memo (optimized)
  │         ├── best: (select G4="[ordering: +1,-2]" G5)
  │         └── cost: 1070.00
  ├── G3: (projections G6 a.x a.y)
- ├── G4: (scan a,cols=(1,2))
+ ├── G4: (scan a,cols=(1,2)) (scan a,rev,cols=(1,2))
  │    ├── ""
  │    │    ├── best: (scan a,cols=(1,2))
  │    │    └── cost: 1060.00
@@ -231,7 +231,7 @@ memo (optimized)
  │         ├── best: (sort G2)
  │         └── cost: 1087.55
  ├── G3: (projections a.y a.z)
- ├── G4: (scan a,cols=(1-3))
+ ├── G4: (scan a,cols=(1-3)) (scan a,rev,cols=(1-3))
  │    ├── ""
  │    │    ├── best: (scan a,cols=(1-3))
  │    │    └── cost: 1070.00
@@ -265,7 +265,7 @@ memo (optimized)
  │    └── "[presentation: Tree:5,Field:8,Description:9,Columns:10,Ordering:11]"
  │         ├── best: (explain G2="[presentation: x:1,y:2,z:3,s:4] [ordering: +2]" [presentation: x:1,y:2,z:3,s:4] [ordering: +2])
  │         └── cost: 1179.66
- └── G2: (scan a)
+ └── G2: (scan a) (scan a,rev)
       ├── ""
       │    ├── best: (scan a)
       │    └── cost: 1080.00
@@ -288,7 +288,7 @@ memo (optimized)
  │    └── ""
  │         ├── best: (row-number G2)
  │         └── cost: 1050.00
- └── G2: (scan a,cols=(2))
+ └── G2: (scan a,cols=(2)) (scan a,rev,cols=(2))
       └── ""
            ├── best: (scan a,cols=(2))
            └── cost: 1050.00
@@ -309,7 +309,7 @@ memo (optimized)
  │         ├── best: (row-number G4)
  │         └── cost: 1050.00
  ├── G3: (projections G5 a.y)
- ├── G4: (scan a,cols=(2))
+ ├── G4: (scan a,cols=(2)) (scan a,rev,cols=(2))
  │    └── ""
  │         ├── best: (scan a,cols=(2))
  │         └── cost: 1050.00
@@ -327,7 +327,7 @@ memo (optimized)
  │    └── ""
  │         ├── best: (row-number G2)
  │         └── cost: 1050.00
- └── G2: (scan a,cols=(2))
+ └── G2: (scan a,cols=(2)) (scan a,rev,cols=(2))
       └── ""
            ├── best: (scan a,cols=(2))
            └── cost: 1050.00
@@ -343,7 +343,7 @@ memo (optimized)
  │    └── ""
  │         ├── best: (row-number G2="[ordering: +2]" ordering=+2)
  │         └── cost: 1149.66
- └── G2: (scan a,cols=(2))
+ └── G2: (scan a,cols=(2)) (scan a,rev,cols=(2))
       ├── ""
       │    ├── best: (scan a,cols=(2))
       │    └── cost: 1050.00
@@ -362,7 +362,7 @@ memo (optimized)
  │    └── ""
  │         ├── best: (row-number G2="[ordering: +2]" ordering=+2)
  │         └── cost: 1149.66
- └── G2: (scan a,cols=(2))
+ └── G2: (scan a,cols=(2)) (scan a,rev,cols=(2))
       ├── ""
       │    ├── best: (scan a,cols=(2))
       │    └── cost: 1050.00
@@ -381,7 +381,7 @@ memo (optimized)
  │    └── ""
  │         ├── best: (row-number G2)
  │         └── cost: 1050.00
- └── G2: (scan a,cols=(2))
+ └── G2: (scan a,cols=(2)) (scan a,rev,cols=(2))
       └── ""
            ├── best: (scan a,cols=(2))
            └── cost: 1050.00