… pkg/sql/opt/exec/execbuilder/testdata/select_index
… pkg/sql/opt/xform/testdata/rules/combo
… pkg/sql/opt/xform/testdata/rules/join
… pkg/sql/opt/xform/testdata/rules/limit
… pkg/sql/opt/xform/testdata/rules/scan
… pkg/sql/opt/xform/testdata/rules/select