The optimizer does not generate plans which include reverse scans.
This PR adds support logical support for reverse scans.

Release note: None