@rytaft 2018-07-02
[nit] fix typo in commit message (support logical support)

Looks good to me, but I'd wait for Radu to comment on that strange DistSQL plan...

---

Reviewed 12 of 12 files at r1, 1 of 1 files at r2.
Review status: :shipit: complete! 0 of 0 LGTMs obtained

---

*[pkg/sql/opt/memo/memo_format.go, line 280 at r1](https://reviewable.io/reviews/cockroachdb/cockroach/27110#-LGRwC2w2FFV29f3uddA:-LGRwC2w2FFV29f3uddB:b7spfx) ([raw file](https://github.com/cockroachdb/cockroach/blob/c94e6392ebfdea6f72032048c81ab859aba4a2dd/pkg/sql/opt/memo/memo_format.go#L280)):*
> ```Go
> 			fmt.Fprintf(f.buf, " %s@%s", tab.TabName(), tab.Index(t.Index).IdxName())
> 		}
> 		fmt.Fprintf(f.buf, ",rev=%t", t.Reverse)
> ```

To avoid cluttering the test files, I'd only include this if `t.Reverse` is true.

---

*[pkg/sql/opt/memo/private_defs.go, line 73 at r1](https://reviewable.io/reviews/cockroachdb/cockroach/27110#-LGRu13F73ilD99WCPjz:-LGRu13F73ilD99WCPk-:b-tcnkom) ([raw file](https://github.com/cockroachdb/cockroach/blob/c94e6392ebfdea6f72032048c81ab859aba4a2dd/pkg/sql/opt/memo/private_defs.go#L73)):*
> ```Go
> // ScanOpDef defines the value of the Def private field of the Scan operator.
> type ScanOpDef struct {
> ```

[nit] extra line

---

*[pkg/sql/opt/memo/private_defs.go, line 114 at r1](https://reviewable.io/reviews/cockroachdb/cockroach/27110#-LGRuo5233ksVFm8-FAZ:-LGRuHPu2mBfUpkYeUcb:b-s0x1te) ([raw file](https://github.com/cockroachdb/cockroach/blob/c94e6392ebfdea6f72032048c81ab859aba4a2dd/pkg/sql/opt/memo/private_defs.go#L114)):*
> ```Go
> 			ordering.AppendCol(colID, !indexCol.Descending)
> 		} else {
> 			ordering.AppendCol(colID, indexCol.Descending)
> ```

You could do something like `ordering.AppendCol(colID, indexCol.Descending != s.Reverse)`

But I think your approach might actually be better since it's very clearly correct. Up to you....

---

*[pkg/sql/opt/memo/private_defs.go, line 115 at r1](https://reviewable.io/reviews/cockroachdb/cockroach/27110#-LGRuPbY3HJ0zbOZdK1N:-LGRuPbY3HJ0zbOZdK1O:b-tcnkom) ([raw file](https://github.com/cockroachdb/cockroach/blob/c94e6392ebfdea6f72032048c81ab859aba4a2dd/pkg/sql/opt/memo/private_defs.go#L115)):*
> ```Go
> 		} else {
> 			ordering.AppendCol(colID, indexCol.Descending)
> ```

[nit] extra line

---


*Comments from [Reviewable](https://reviewable.io/reviews/cockroachdb/cockroach/27110#-:-LGRsjHJDooYnAwngsQM:b-259fnw)*
<!-- Sent from Reviewable.io -->