@madhavsuresh 2018-07-02
TFTR! Fixed the typo, sounds good on waiting to discuss the DistSQL plan. 

---

Review status: :shipit: complete! 0 of 0 LGTMs obtained

---

*[pkg/sql/opt/memo/memo_format.go, line 280 at r1](https://reviewable.io/reviews/cockroachdb/cockroach/27110#-LGRwC2w2FFV29f3uddA:-LGS-YBG0BfXr53-OWgI:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/c94e6392ebfdea6f72032048c81ab859aba4a2dd/pkg/sql/opt/memo/memo_format.go#L280)):*
<details><summary><i>Previously, rytaft wrote…</i></summary><blockquote>

To avoid cluttering the test files, I'd only include this if `t.Reverse` is true.
</blockquote></details>

Done.

---

*[pkg/sql/opt/memo/private_defs.go, line 73 at r1](https://reviewable.io/reviews/cockroachdb/cockroach/27110#-LGRu13F73ilD99WCPjz:-LGS-Zn7554GMIuYCt6G:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/c94e6392ebfdea6f72032048c81ab859aba4a2dd/pkg/sql/opt/memo/private_defs.go#L73)):*
<details><summary><i>Previously, rytaft wrote…</i></summary><blockquote>

[nit] extra line
</blockquote></details>

Done.

---

*[pkg/sql/opt/memo/private_defs.go, line 114 at r1](https://reviewable.io/reviews/cockroachdb/cockroach/27110#-LGRuo5233ksVFm8-FAZ:-LGS-c-p2Dp1kXgp5iL8:b-11xfk4) ([raw file](https://github.com/cockroachdb/cockroach/blob/c94e6392ebfdea6f72032048c81ab859aba4a2dd/pkg/sql/opt/memo/private_defs.go#L114)):*
<details><summary><i>Previously, rytaft wrote…</i></summary><blockquote>

You could do something like `ordering.AppendCol(colID, indexCol.Descending != s.Reverse)`

But I think your approach might actually be better since it's very clearly correct. Up to you....
</blockquote></details>

Yea I see what you're saying. However, at the risk of being too verbose, the current if statement is easier for me to understand. 

---

*[pkg/sql/opt/memo/private_defs.go, line 115 at r1](https://reviewable.io/reviews/cockroachdb/cockroach/27110#-LGRuPbY3HJ0zbOZdK1N:-LGS-b1h7MvaIB5rpSSj:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/c94e6392ebfdea6f72032048c81ab859aba4a2dd/pkg/sql/opt/memo/private_defs.go#L115)):*
<details><summary><i>Previously, rytaft wrote…</i></summary><blockquote>

[nit] extra line
</blockquote></details>

Done.

---


*Comments from [Reviewable](https://reviewable.io/reviews/cockroachdb/cockroach/27110#-:-LGS03A768RahdciSsjz:b-q5gxjz)*
<!-- Sent from Reviewable.io -->