@justinj 2018-07-03
We probably want to cost this as a bit more expensive than a normal scan, I think? I remember hearing anecdotally that they're ~3-5x slower than forward scans in RocksDB. That's definitely a concern that can be pushed down the road though (maybe consult Peter about the relative rate, I remember him talking about it a while back).

---

Review status: :shipit: complete! 0 of 0 LGTMs obtained

---

*[pkg/sql/opt/xform/custom_funcs.go, line 86 at r3](https://reviewable.io/reviews/cockroachdb/cockroach/27110#-LGSU53X-q5rOcW085zk:-LGSU53X-q5rOcW085zl:b37r9gp) ([raw file](https://github.com/cockroachdb/cockroach/blob/384f21ba3b3e0bd922ae9defd154bca371249e0d/pkg/sql/opt/xform/custom_funcs.go#L86)):*
> ```Go
> 		// then construct a new Scan operator using that index.
> 		if scanOpDef.Cols.SubsetOf(indexCols) {
> 			newDef := &memo.ScanOpDef{Table: scanOpDef.Table, Index: i, Cols: scanOpDef.Cols, Reverse: false}
> ```

up to you, but since `false` is the zero-value of `bool`, you can omit the `Reverse: false`, which still seems reasonably semantic to me, since non-reverse is the "default".

---

*[pkg/sql/opt/xform/custom_funcs.go, line 109 at r3](https://reviewable.io/reviews/cockroachdb/cockroach/27110#-LGSUQUr1Pln94A6eTpk:-LGSUQUr1Pln94A6eTpl:b-2guby3) ([raw file](https://github.com/cockroachdb/cockroach/blob/384f21ba3b3e0bd922ae9defd154bca371249e0d/pkg/sql/opt/xform/custom_funcs.go#L109)):*
> ```Go
> 				Index:   i,
> 				Cols:    scanCols,
> 				Reverse: false,
> ```

same here

---

*[pkg/sql/opt/xform/testdata/rules/scan, line 36 at r3](https://reviewable.io/reviews/cockroachdb/cockroach/27110#-LGSSw4b37Y3dbzfmkJw:-LGSSw4b37Y3dbzfmkJx:bawn3xr) ([raw file](https://github.com/cockroachdb/cockroach/blob/384f21ba3b3e0bd922ae9defd154bca371249e0d/pkg/sql/opt/xform/testdata/rules/scan#L36)):*
> ```
> # --------------------------------------------------
> 
> # Revscan won't be used here because there is no index with f
> ```

If there was an `f DESC` index, wouldn't that be serviced by a forward scan?

---


*Comments from [Reviewable](https://reviewable.io/reviews/cockroachdb/cockroach/27110#-:-LGSQnxwAsbwLo10FzYU:b-uyocox)*
<!-- Sent from Reviewable.io -->