@petermattis 2018-07-03
> We probably want to cost this as a bit more expensive than a normal scan, I think? I remember hearing anecdotally that they're ~3-5x slower than forward scans in RocksDB. That's definitely a concern that can be pushed down the road though (maybe consult Peter about the relative rate, I remember him talking about it a while back).

The performance diff depends on the number of historic versions. Assuming 2-3x slower seems reasonable (below, old is forward scan and new is reverse scan).

```
name                                                   old time/op    new time/op    delta
MVCCScan_RocksDB/rows=1000/versions=1/valueSize=8-8       203µs ± 1%     341µs ± 2%   +68.06%  (p=0.000 n=8+10)
MVCCScan_RocksDB/rows=1000/versions=2/valueSize=8-8       264µs ± 2%     742µs ± 4%  +181.42%  (p=0.000 n=10+10)
MVCCScan_RocksDB/rows=1000/versions=10/valueSize=8-8      700µs ± 2%    2346µs ± 2%  +235.15%  (p=0.000 n=10+10)
MVCCScan_RocksDB/rows=1000/versions=100/valueSize=8-8    4.54ms ± 1%    8.82ms ± 3%   +94.32%  (p=0.000 n=9+10)
```

---

Review status: :shipit: complete! 0 of 0 LGTMs obtained

---



*Comments from [Reviewable](https://reviewable.io/reviews/cockroachdb/cockroach/27110#-:-LGSrqeN2cVzsDuUZ23a:b-xhyw9x)*
<!-- Sent from Reviewable.io -->